+++ example/lib/main.dart
@@ -19,10 +19,7 @@ void main() async {
     adIdManager,
     adMobAdRequest: const AdRequest(),
     admobConfiguration: RequestConfiguration(testDeviceIds: []),
-    fbTestingId: '73f92d66-f8f6-4978-999f-b5e0dd62275a',
-    fbTestMode: true,
     showAdBadge: Platform.isIOS,
-    fbiOSAdvertiserTrackingEnabled: true,
   );
 
   runApp(const MyApp());
@@ -90,11 +87,7 @@ class _CountryListScreenState extends State<CountryListScreen> {
                 networkName: 'Jit Admob Interstitial',
                 onTap: () => _showJitInterstitial(context),
               ),
-              AdButton(
-                networkName: 'Facebook Interstitial',
-                onTap: () =>
-                    _showAd(AdNetwork.facebook, AdUnitType.interstitial),
-              ),
+
               AdButton(
                 networkName: 'Available Interstitial',
                 onTap: () => _showAvailableAd(AdUnitType.interstitial),
@@ -115,17 +108,12 @@ class _CountryListScreenState extends State<CountryListScreen> {
                 networkName: 'Jit Admob Rewarded',
                 onTap: () => _showJitRewarded(),
               ),
-              AdButton(
-                networkName: 'Facebook Rewarded',
-                onTap: () => _showAd(AdNetwork.facebook, AdUnitType.rewarded),
-              ),
+
               AdButton(
                 networkName: 'Available Rewarded',
                 onTap: () => _showAvailableAd(AdUnitType.rewarded),
               ),
-              const EasySmartBannerAd(
-                priorityAdNetworks: [AdNetwork.facebook, AdNetwork.admob],
-              ),
+              const EasySmartBannerAd(priorityAdNetworks: [AdNetwork.admob]),
             ],
           ),
         ),