+++ lib/src/enums/ad_network.dart
@@ -1,4 +1,4 @@
-enum AdNetwork { any, admob, facebook }
+enum AdNetwork { any, admob }
 
 extension AdNetworkExtension on AdNetwork {
   String get value => name;